requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
python-dotenv==1.0.1
openai==1.30.1
azure-cognitiveservices-speech==1.37.0
azure-ai-vision-imageanalysis==1.0.0
azure-ai-textanalytics==5.3.0
aiohttp==3.9.5
aiofiles==23.2.1
python-multipart==0.0.9
httpx==0.27.0